Move column configs to nearby places where being used
