+++ src/components/button/ExcelDownloadCButton.js
@@ -1,11 +1,15 @@
 import { CButton } from '@coreui/react-pro';
 
-const ExcelDownloadCButton = ({ downloadFunction, searchFormData }) => {
+const ExcelDownloadCButton = ({ downloadFunction, searchFormData, hasSearchResults }) => {
   const handleDownload = async () => {
     await downloadFunction(searchFormData);
   };
 
-  return <CButton onClick={handleDownload}>엑셀 다운로드</CButton>;
+  return (
+    <CButton onClick={handleDownload} disabled={!hasSearchResults}>
+      엑셀 다운로드
+    </CButton>
+  );
 };
 
 export default ExcelDownloadCButton;
