+++ src/views/pages/document-collection/documentCollectionColumnConfig.js
